# Conflicts:
#	modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/SqlQueryProcessor.java
#	modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/ddl/DdlCommandHandler.java
#	modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/exec/ddl/DdlCommandHandlerWrapper.java
#	modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/schema/CatalogSqlSchemaManager.java
#	modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/schema/SqlSchemaManagerImpl.java
#	modules/table/src/main/java/org/apache/ignite/internal/table/KeyValueBinaryViewImpl.java
#	modules/table/src/main/java/org/apache/ignite/internal/table/KeyValueViewImpl.java
#	modules/table/src/main/java/org/apache/ignite/internal/table/RecordBinaryViewImpl.java
#	modules/table/src/main/java/org/apache/ignite/internal/table/RecordViewImpl.java
#	modules/table/src/main/java/org/apache/ignite/internal/table/distributed/TableManager.java

+++ modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/message/QueryStartRequest.java
@@ -58,5 +58,5 @@ public interface QueryStartRequest extends ExecutionContextAwareMessage {
     /**
      * Return last schema version, just a stub, need to be removed after IGNITE-18733.
      */
-    long schemaVersion();
+    int schemaVersion();
 }
